simplify generate tags, rename smk-tags to tags
