+++ gruntfile.js
@@ -1,8 +1,8 @@
 module.exports = function( grunt ) {
 
-    var path = require( "path" );
-
     require( 'load-grunt-tasks' )( grunt )
+
+    grunt.loadTasks( 'tasks' )
 
     var jshintStylish = require( 'jshint-stylish' )
     jshintStylish.reporter = ( function ( inner ) {
@@ -26,31 +26,6 @@ module.exports = function( grunt ) {
                 grunt.log.writeln( srcpath + ': ' + m );
                 return grunt.template.process( m );
             } )
-        },
-
-        // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-
-        connectOption: {
-            http: {
-                options: {
-                    protocol: 'http',
-                    hostname: '*',
-                    port: 8888,
-                    base: '<%= buildPath %>',
-                    livereload: true,
-                    // debug: true
-                }
-            },
-            https: {
-                options: {
-                    protocol: 'https',
-                    hostname: '*',
-                    port: 8443,
-                    base: '<%= buildPath %>',
-                    livereload: true,
-                    // debug: true
-                }
-            }
         },
 
         // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
@@ -190,26 +165,35 @@ module.exports = function( grunt ) {
 
         // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
-        // filelist: {
-        //     configs: {
-        //         files: [
-        //             {
-        //                 cwd: 'src/main/test/config',
-        //                 src: [
-        //                     '*json',
-        //                 ],
-        //                 dest: '<%= buildPath %>/test/configs.json'
-        //             }
-        //         ]
-        //     },
-        // },
+        connectOption: {
+            http: {
+                options: {
+                    protocol: 'http',
+                    hostname: '*',
+                    port: 8888,
+                    base: '<%= buildPath %>',
+                    livereload: true,
+                    // debug: true
+                }
+            },
+            https: {
+                options: {
+                    protocol: 'https',
+                    hostname: '*',
+                    port: 8443,
+                    base: '<%= buildPath %>',
+                    livereload: true,
+                    // debug: true
+                }
+            }
+        },
 
         // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
         watch: {
             options: {
+                debounceDelay: 2000,
                 livereload: {
-                    // debounceDelay: 5000,
                     host:   '<%= serverHost %>',
                     key:    grunt.file.read( 'node_modules/grunt-contrib-connect/tasks/certs/server.key' ),
                     cert:   grunt.file.read( 'node_modules/grunt-contrib-connect/tasks/certs/server.crt' )
@@ -227,23 +211,11 @@ module.exports = function( grunt ) {
             test: {
                 files: [ '<%= examplePath %>/**' ],
                 tasks: [ 'build-examples', 'build-example-data' ]
-            },
+            }
         }
 
     } )
 
-    grunt.registerTask( 'gen-tags', function () {
-        // seems to be only way to clear require cache
-        for ( var key in require.cache )
-            delete require.cache[ key ]
-
-        var tags = require( './smk-tags' )
-        grunt.log.write( 'Generating tags...' )
-        var tagData = tags.gen()
-        grunt.log.ok()
-
-        grunt.config( 'tag', JSON.parse( JSON.stringify( tagData ) ) )
-    } )
 
     grunt.registerTask( 'deploy', 'set deploy dir', function ( dir ) {
         grunt.config( 'deployPath', dir )
@@ -294,35 +266,6 @@ module.exports = function( grunt ) {
             'connect',
             'watch'
         )
-    } )
-
-    grunt.registerMultiTask( 'filelist', 'Writes JSON blobs containing names of the matched files to sub-keys for destination in a config setting', function ( setting ) {
-        var out = {};
-        this.files.forEach( function ( f ) {
-            var cwd = f.cwd || '';
-
-            var dest = path.basename( f.dest, path.extname( f.dest ) )
-
-            var list = f.src.map( function ( filename ) {
-                var s = path.join( cwd, filename )
-                if ( !grunt.file.isFile( s ) ) return;
-
-                grunt.log.writeln( dest + ': ' + filename )
-                return {
-                    // name: path.basename( s ),
-                    path: filename,
-                }
-            } ).filter( function ( e ) { return !!e } )
-
-            out[ dest ] = list
-        } )
-
-        if ( setting )
-            grunt.config( setting, jsonOut( out ) )
-
-        function jsonOut( obj ) {
-            return JSON.stringify( obj )
-        }
     } )
 
     grunt.registerTask( 'build-info', function () {
